Rename list_keys_result KeyVault header to list_keys_responses to match PagedResponse naming pattern (#2592)
* rename header

* format
